review-app/requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
openai
boto3  
requests  
minio